review: changes part 2
- combined the implementation back to one file,
  different/not supported features on the older chip
  are  guarded by feature gates.
- IRQs adjusted to handle differences between chips.
- dropped unrelated changes
