+++ rnfc-st25r39/src/impls/mod.rs
@@ -8,6 +8,8 @@ pub mod lib_st25r3911b;
 #[cfg(feature = "st25r3911b")]
 pub use lib_st25r3911b::*;
 
+pub mod lib;
+
 // TODO: check wup and other wake-up registers in st25r3911b
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
